refactor(frontend): canonicalize SearchableSelect (#4359)
- Move API-backed entity SearchableSelect into components/Shared

- Keep FormSubmission/SearchableSelect as a compatibility re-export

- Rename Fuse-based version to LocalSearchableSelect

- Break circular import by importing EntityFormSurface directly in QuickCreateModal
